+++ online/src/ui/viewer/store.js
@@ -165,90 +165,6 @@ const reducer = ( state = initialState, event ) =>
   }
 };
 
-const serializeLighting = ( lighting, doc ) =>
-{
-  const getRGB = color =>{
-    const [ r, g, b ] = [ parseInt( color.substr(1, 2), 16 ), parseInt( color.substr(3, 2), 16 ), parseInt( color.substr(5, 2), 16 ) ];
-    return r + ',' + g + ',' + b;
-  }
-  const { ambientColor, backgroundColor, directionalLights } = lighting;
-  const sceneModel = doc .createElement( "sceneModel" );
-  sceneModel .setAttribute( "ambientLight", getRGB( ambientColor ) );
-  sceneModel .setAttribute( "background", getRGB( backgroundColor ) );
-  for (const dlight of directionalLights) {
-    const { color, direction } = dlight;
-    const child = doc .createElement( "directionalLight" );
-    child .setAttribute( "color", getRGB( color ) );
-    child .setAttribute( "x", String( direction[ 0 ] ) );
-    child .setAttribute( "y", String( direction[ 1 ] ) );
-    child .setAttribute( "z", String( direction[ 2 ] ) );
-    sceneModel .appendChild( child );
-  }
-  return sceneModel;
-}
-
-const serializeCamera = ( camera, doc ) =>
-{
-  const { near, far, width, distance, perspective, lookAt, lookDir, up } = camera;
-  const viewing = doc .createElement( "Viewing" );
-  const viewmodel = doc .createElement( "ViewModel" );
-  viewing .appendChild( viewmodel );
-  viewmodel .setAttribute( "near", String( near ) );
-  viewmodel .setAttribute( "far", String( far ) );
-  viewmodel .setAttribute( "width", String( width ) );
-  viewmodel .setAttribute( "distance", String( distance ) );
-  viewmodel .setAttribute( "stereoAngle", "0.0" );
-  viewmodel .setAttribute( "parallel", String( !perspective ) );
-  {
-      const child = doc .createElement( "LookAtPoint" );
-      child .setAttribute( "x", String( lookAt[ 0 ] ) );
-      child .setAttribute( "y", String( lookAt[ 1 ] ) );
-      child .setAttribute( "z", String( lookAt[ 2 ] ) );
-      viewmodel .appendChild( child );
-  }
-  {
-      const child = doc .createElement( "UpDirection" );
-      child .setAttribute( "x", String( up[ 0 ] ) );
-      child .setAttribute( "y", String( up[ 1 ] ) );
-      child .setAttribute( "z", String( up[ 2 ] ) );
-      viewmodel .appendChild( child );
-  }
-  {
-      const child = doc .createElement( "LookDirection" );
-      child .setAttribute( "x", String( lookDir[ 0 ] ) );
-      child .setAttribute( "y", String( lookDir[ 1 ] ) );
-      child .setAttribute( "z", String( lookDir[ 2 ] ) );
-      viewmodel .appendChild( child );
-  }
-  return viewing;
-}
-
-export const serializeVZomeXml = ( xmlStr, camera, lighting ) =>
-{
-  const parser = new DOMParser();
-  const doc = parser .parseFromString( xmlStr, "application/xml" );
-
-  const rootElement = doc .documentElement;
-  const field = rootElement .getAttribute( 'field' );
-
-  const children = rootElement .childNodes;
-
-  const newRoot = document .createElementNS( 'http://xml.vzome.com/vZome/4.0.0/', 'vzome:vZome' );
-  newRoot .setAttribute( 'field', field );
-  newRoot .setAttribute( 'edition', 'online' );
-  newRoot .setAttribute( 'version', '1.0' );
-  newRoot .setAttribute( 'buildNumber', REVISION );
-  newRoot .replaceChildren( ...children );
-
-  newRoot .appendChild( serializeLighting( lighting, doc ) );
-  newRoot .appendChild( serializeCamera( camera, doc ) );
-
-  rootElement .replaceWith( newRoot );
-
-  const serializer = new XMLSerializer();
-  return serializer .serializeToString( doc );
-}
-
 const branchSelectionBlocks = node =>
 {
   if ( node.children && node.children.length > 1 ) {
